app/requirements.txt
asgiref==3.2.10
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.11.0
django-environ==0.4.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-extra-fields==3.2.1
gunicorn==20.1.0
idna==3.3
postgis==1.0.4
psycopg2==2.9.3
psycopg2-binary==2.9.1
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
